+++ src/core/Core.js
@@ -3,6 +3,7 @@ import Translator from '../core/Translator'
 import prettyBytes from 'pretty-bytes'
 import yo from 'yo-yo'
 import ee from 'events'
+import deepFreeze from 'deep-freeze-strict'
 import UppySocket from './UppySocket'
 
 /**
@@ -33,6 +34,8 @@ export default class Core {
 
     this.translator = new Translator({locales: this.opts.locales})
     this.i18n = this.translator.translate.bind(this.translator)
+    this.getState = this.getState.bind(this)
+    this.updateMeta = this.updateMeta.bind(this)
     this.initSocket = this.initSocket.bind(this)
 
     this.emitter = new ee.EventEmitter()
@@ -53,10 +56,10 @@ export default class Core {
    * Iterate on all plugins and run `update` on them. Called each time when state changes
    *
    */
-  updateAll () {
+  updateAll (state) {
     Object.keys(this.plugins).forEach((pluginType) => {
       this.plugins[pluginType].forEach((plugin) => {
-        plugin.update()
+        plugin.update(this.state)
       })
     })
   }
@@ -66,11 +69,15 @@ export default class Core {
    *
    * @param {newState} object
    */
-  setState (newState) {
-    this.log('Setting state to: ')
+  setState (stateUpdate) {
+    const newState = Object.assign({}, this.state, stateUpdate)
+    this.emitter.emit('state-update', this.state, newState, stateUpdate)
+
+    this.state = newState
+    this.updateAll(this.state)
+
+    this.log('Updating state with: ')
     this.log(newState)
-    this.state = Object.assign({}, this.state, newState)
-    this.updateAll()
   }
 
   /**
@@ -79,15 +86,28 @@ export default class Core {
    *
    */
   getState () {
-    return this.state
+    return deepFreeze(this.state)
   }
 
-  addMeta (meta, fileID) {
+  updateMeta (data, fileID) {
+    // if fileID is not specified, set meta data to all files
     if (typeof fileID === 'undefined') {
-      const updatedFiles = Object.assign({}, this.state.files)
-      for (let file in updatedFiles) {
-        updatedFiles[file].meta = meta
-      }
+      const updatedFiles = Object.assign({}, this.getState().files)
+      Object.keys(updatedFiles).forEach((file) => {
+        const newMeta = Object.assign({}, updatedFiles[file].meta, data)
+        updatedFiles[file] = Object.assign({}, updatedFiles[file], {
+          meta: newMeta
+        })
+      })
+      this.setState({files: updatedFiles})
+
+    // if fileID is specified, set meta to that file
+    } else {
+      const updatedFiles = Object.assign({}, this.getState().files)
+      const newMeta = Object.assign({}, updatedFiles[fileID].meta, data)
+      updatedFiles[fileID] = Object.assign({}, updatedFiles[fileID], {
+        meta: newMeta
+      })
       this.setState({files: updatedFiles})
     }
   }
@@ -107,6 +127,9 @@ export default class Core {
       id: fileID,
       name: file.name,
       extension: fileExtension,
+      meta: {
+        name: file.name
+      },
       type: {
         general: fileTypeGeneral,
         specific: fileTypeSpecific
@@ -130,8 +153,12 @@ export default class Core {
         const newImageHeight = Utils.getProportionalImageHeight(imgEl, newImageWidth)
         const resizedImgSrc = Utils.resizeImage(imgEl, newImageWidth, newImageHeight)
 
-        const updatedFiles = Object.assign({}, this.state.files)
-        updatedFiles[fileID].previewEl = yo`<img alt="${file.name}" src="${resizedImgSrc}">`
+        const updatedFiles = Object.assign({}, this.getState().files)
+        const updatedFile = Object.assign({}, updatedFiles[fileID], {
+          previewEl: yo`<img alt="${file.name}" src="${resizedImgSrc}">`,
+          preview: resizedImgSrc
+        })
+        updatedFiles[fileID] = updatedFile
         this.setState({files: updatedFiles})
       })
     }
@@ -164,8 +191,11 @@ export default class Core {
       percentage = Math.round(percentage)
 
       const updatedFiles = Object.assign({}, this.state.files)
-      updatedFiles[data.id].progress = percentage
-      updatedFiles[data.id].uploadedSize = data.bytesUploaded ? prettyBytes(data.bytesUploaded) : '?'
+      const updatedFile = Object.assign({}, updatedFiles[data.id], {
+        progress: percentage,
+        uploadedSize: data.bytesUploaded ? prettyBytes(data.bytesUploaded) : '?'
+      })
+      updatedFiles[data.id] = updatedFile
 
       const inProgress = Object.keys(updatedFiles).map((file) => {
         return file.progress !== 0
@@ -187,12 +217,10 @@ export default class Core {
       })
     })
 
-    // `upload-success` adds successfully uploaded file to `state.uploadedFiles`
-    // and fires `remove-file` to remove it from `state.files`
     this.emitter.on('upload-success', (file) => {
-      const updatedFiles = Object.assign({}, this.state.files)
-      updatedFiles[file.id] = file
-      this.setState({files: updatedFiles})
+      // const updatedFiles = Object.assign({}, this.state.files)
+      // updatedFiles[file.id] = file
+      // this.setState({files: updatedFiles})
     })
   }
 
@@ -273,7 +301,7 @@ export default class Core {
     if (msg === `${msg}`) {
       console.log(`LOG: ${msg}`)
     } else {
-      console.log('LOG↓')
+      // console.log('LOG↓')
       console.dir(msg)
     }
     global.uppyLog = global.uppyLog + '\n' + 'DEBUG LOG: ' + msg